Don't need a separate /index.html entry in routes.
Have the offline index.html refresh to URL "/" since the web server shows
index.html for the "/" root URL if Passenger/Rails is down.  If Passenger
is up, it runs the application at that URL.  Only worry is for users in
the middle of doing something and ending up at a dead page when submitting.