Work in progress.  Trying to get new database design working.
Some problems with lack of polymorphism in single table inheritance
when searching through an intermediate table (the links) and Rails only
uses a specific class of intermediate link, ignoring sub/super classes.

Also now thinking that a new table for references to links to delete and giving
permissions will be needed, rather than cramming all that into LinkBase with
not quite working subclasses of LinkBase for links to ledger table objects vs
links to link objects.