Handle case where original_id is nil, start testing LedgerFullGroup.
Turns out data generated by Fixtures has a nil original_id.  Yes
we can specify it, and set the id explicitly, but record id and
database idea of it are different and fetching records by id
stops working.  So put up with a nil original_id in the utility
functions for finding the original record or latest version, etc.