WIP - Adding reply links, so you can post a reply to a post.
There is some complexity with editing reply messages before they are
saved.  Can carry around the link to the original message and posting
group as parameter based attribute records that only exist in memory
and in forms.  But it has problems being saved if you are editing an
existing message, since the edited message is actually a new record
and that confuses ActiveRecord's param to record facility.  Working
on that...