Add checkpoint fields to the LedgerBase record.
Useful for easier point recalculation, since we save the values after the weekly ceremony, including bonus points and fading.  Then we can recalculate from the checkpoint to the present time if needed, rather than going all the way back to the beginning of time, which would be a big job for people with thousands of posts.