WIP Thinking about checkpointing rating points, maybe wrong.
The trouble is the weekly awards of allowance points may be affected by garbage collecting records.  But we also want to recalculate things if links are deleted or undeleted or unapproved, etc.  Saving a checkpoint sounded good, but in retrospect is useless if we have current values affected by deletion of something in the distant past before the checkpoint.  Perhaps just need to save the weekly allowance accumulated total for each user and say that past allowances aren't affected by recent deletions.  For other objects, we only really delete them when the net value of their points has gotten close to zero, so removing them won't suddenly give the user a lot of extra spent points back - they'll get back a faded close to zero amount, so we can ignore that slight distortion and do a simple recalculation of everything.

So, if deleting something in the current week, just update the current points.  If deleting something older, mark it as needing a full recalculation.  For Users, save their faded accumulated allowance each week (reflecting current deleted status of bonus point records) and add it to the full recalculation.