More reputation point calculations and test code.
More work on updating reputation points over time, now handling changes in approvals and deletion of links that affects the points applied to LedgerBase objects.  Tests added.

When marking records as deleted or approved, make note of the case where nothing changes, and revert the creation of the LedgerDelete or LedgerApprove record.