Switch to using string of ID numbers for path loop detection.
List of all descendants of a post is starting to work.  Using a
string instead of an ARRAY of ID numbers so that Sqlite3 can
work with the same SQL query as we use in production.  Need to
fix up descendants listing to use better indenting and colour bars
too if possible (bunch of narrow DIVs with background colours?).